moved slow-travel-mini-festival and festival banner
