+++ src/pages/index.svelte
@@ -3,16 +3,16 @@
   import smoothscroll from 'smoothscroll-polyfill';
   import routes from '@/routes';
   import CollapsibleGroup from '../components/CollapsibleGroup.svelte';
-  import { Button, Card, Icon } from '../components/UI';
-  import { getArrayFromLocale, transKeyExists, getCookie, setCookie } from '@/util';
+  import { Button } from '../components/UI';
+  import SlowTravelMiniFestival from '../components/Temporary/SlowTravelMiniFestival.svelte';
+  import FestivalBanner from '../components/Temporary/FestivalBanner.svelte';
+  import { getArrayFromLocale, transKeyExists } from '@/util';
   import { user } from '@/stores/auth';
 
-  import { crossIcon } from '@/images/icons';
   import Logo from '../images/logo.svg';
   import welcomeMap from '../images/welcome-map.svg';
   import ArrowDown from '../images/arrow-down.svg';
   import OKLogo from '../images/ok_logo.svg';
-  import VGCLogo from '../images/vgc_logo.svg';
   import natuurpuntLogo from '../images/natuurpunt_logo.svg';
   import groteRoutePadenLogo from '../images/groteroutepaden-logo.svg';
   import GRSentiers from '../images/les-sentiers-de-grande-randonnee-logo.svg';
@@ -30,15 +30,6 @@
     window.scroll({ top: topOfStepsSection, behavior: 'smooth' });
   }
 
-  let festivalBannerShown = !getCookie('festival-banner-dismissed');
-  const closeBanner = () => {
-    const date = new Date();
-    // one year
-    date.setTime(date.getTime() + 365 * 86400000); //24 * 60 * 60 * 1000
-    setCookie('festival-banner-dismissed', true, { expires: date.toGMTString() });
-    festivalBannerShown = false;
-  };
-
   smoothscroll.polyfill();
 
   const stepGraphics = [Step1, Step2, Step3];
@@ -50,30 +41,6 @@
 <svelte:head>
   <title>{$_('generics.home')} | Welcome To My Garden</title>
 </svelte:head>
-
-{#if festivalBannerShown}
-  <header class="banner">
-    <div class="banner-content">
-      <p>
-        {@html $_('index.banner-stmf.copy', {
-          values: {
-            stmfLink: `<a class="banner-link" href="#mini-festival">${$_(
-              `index.banner-stmf.link`
-            )}</a>`
-          }
-        })}
-      </p>
-
-      <button
-        on:click={closeBanner}
-        aria-label="Close banner"
-        class="button-container close-banner"
-      >
-        <Icon icon={crossIcon} />
-      </button>
-    </div>
-  </header>
-{/if}
 
 <section class="landing" id="landing">
   <div class="welcome">
@@ -137,105 +104,6 @@
   {/each}
 </section>
 
-<section class="slow-travel-mini-festival" id="mini-festival">
-  <div class="stmf-intro">
-    <h1 class="stmf-intro-title">{$_('stmf.title')}</h1>
-    <p class="stmf-intro-byline">{$_('stmf.description')}</p>
-  </div>
-
-  <div class="stmf-columns">
-    <div class="stmf-column">
-      <div class="stmf-column-header">
-        <h4>{$_('stmf.sections.get-inspired')}</h4>
-      </div>
-      <Card
-        languageAbbreviation="EN"
-        src="images/workshops/the-1001-ways-of-slow-travelling.jpg"
-        date="3 June 2021"
-        time="19:00 - 20:30"
-        title="The 1001 ways of slow travelling"
-        href="https://1001ways.eventbrite.be/?aff=Website"
-      />
-    </div>
-
-    <div class="stmf-column">
-      <div class="stmf-column-header">
-        <h4>{$_('stmf.sections.plan')}</h4>
-      </div>
-      <Card
-        languageAbbreviation="NL"
-        src="/images/workshops/creeer-je-eigen-route-praktische-tips-tools.jpg"
-        date="8 June 2021"
-        time="19:00 - 20:30"
-        title="Creëer je eigen route: praktische tips & tools"
-        href="https://eigenroute.eventbrite.be/?aff=Website"
-      />
-      <Card
-        languageAbbreviation="FR"
-        src="/images/workshops/creer-ton-propre-itineraire-conseils-pratiques-outils.jpg"
-        date="10 June 2021"
-        time="19:00 - 20:30"
-        title="Créer ton propre itinéraire: conseils pratiques & outils"
-        href="https://propreitineraire.eventbrite.be/?aff=Website"
-      />
-      <Card
-        languageAbbreviation="EN"
-        src="/images/workshops/create-your-own-itinerary-practical-tips-tools.jpg"
-        date="15 June 2021"
-        time="19:00 - 20:30"
-        title="Create your own itinerary: practical tips & tools"
-        href="https://ownitinerary.eventbrite.be/?aff=Website"
-      />
-    </div>
-
-    <div class="stmf-column">
-      <div class="stmf-column-header">
-        <h4>{$_('stmf.sections.contribute')}</h4>
-      </div>
-      <Card
-        group="Beginner"
-        languageAbbreviation="EN"
-        src="/images/workshops/collect-data-and-improve-maps-while-slow-travelling-introduction-to-openstreetmap-beginner.jpg"
-        date="17 June 2021"
-        time="19:00 - 20:30"
-        title="Collect data and improve maps while slow travelling: Introduction to OpenStreetMap"
-        href="https://openstreetmap-beginner.eventbrite.be/?aff=Website"
-      />
-      <Card
-        group="Advanced"
-        languageAbbreviation="EN"
-        src="/images/workshops/collect-data-and-improve-maps-while-slow-travelling-introduction-to-openstreetmap.jpg"
-        date="22 June 2021"
-        time="19:00 - 20:30"
-        title="Collect data and improve maps while slow travelling: Deep dive into OpenStreetMap"
-        href="https://openstreetmap-advanced.eventbrite.be/?aff=Website"
-      />
-      <Card
-        languageAbbreviation="EN"
-        src="/images/workshops/opportunities-of-worlds-biggest-encyclopedia-for-slow-travellers-wikipedia.jpg"
-        date="24 June 2021"
-        time="19:00 - 20:30"
-        title="Opportunities of world's biggest encyclopedia for slow travellers: Wikipedia"
-        href="https://wikipedia-workshop.eventbrite.be/?aff=Website"
-      />
-    </div>
-
-    <div class="stmf-column">
-      <div class="stmf-column-header">
-        <h4>{$_('stmf.sections.rethink')}</h4>
-      </div>
-      <Card
-        languageAbbreviation="EN"
-        src="/images/workshops/slow-travelling-the-commons-the-role-of-the-welcome-to-my-garden-community.jpg"
-        date="29 June 2021"
-        time="19:00 - 20:30"
-        title="Slow travelling & the commons: the role of the Welcome To My Garden community"
-        href="https://slowtravel-commons.eventbrite.be/?aff=Website"
-      />
-    </div>
-  </div>
-</section>
-
 <section class="faq" id="faq">
   <div class="card faq-intro">
     <h1 class="heading-underline-center">{$_('index.faq.title')}</h1>
@@ -263,19 +131,19 @@
           </a>
         </div>
         <div class="partner-logo grouteroutepaden-logo">
-          <a href="https://www.groteroutepaden.be/" class="partner-link"  target="_blank">
+          <a href="https://www.groteroutepaden.be/" class="partner-link" target="_blank">
             {@html groteRoutePadenLogo}
           </a>
         </div>
       </div>
       <div>
         <div class="partner-logo ok-logo">
-          <a href="https://be.okfn.org/" class="partner-link"  target="_blank">
+          <a href="https://be.okfn.org/" class="partner-link" target="_blank">
             {@html OKLogo}
           </a>
         </div>
         <div class="partner-logo natuurpunt-logo">
-          <a href="https://www.natuurpunt.be/" class="partner-link"  target="_blank">
+          <a href="https://www.natuurpunt.be/" class="partner-link" target="_blank">
             {@html natuurpuntLogo}
           </a>
         </div>
@@ -367,55 +235,6 @@
     background: var(--color-orange);
     width: 12rem;
     margin-left: -6rem;
-  }
-
-  .banner {
-    position: fixed;
-    width: 100vw;
-    z-index: 100;
-    min-height: 9.2rem;
-    box-shadow: 0px 15px 10px -15px var(--color-black);
-    background-color: var(--color-green);
-    text-align: center;
-    left: 0;
-  }
-
-  .banner-content {
-    width: 100%;
-    display: flex;
-    align-content: center;
-    justify-content: center;
-    padding: 2rem 8rem;
-  }
-
-  .banner p {
-    color: var(--color-white);
-    font-weight: bold;
-    max-width: 80rem;
-    margin: 0 auto;
-    text-align: center;
-  }
-
-  .close-banner {
-    width: 3.6rem;
-    height: 3.6rem;
-    position: absolute;
-    right: 2.2rem;
-    top: calc(50% - 1.8rem);
-    cursor: pointer;
-    z-index: 110;
-  }
-
-  .close-banner :global(svg) {
-    fill: var(--color-white);
-  }
-
-  :global(.banner-link) {
-    text-decoration: underline;
-  }
-
-  :global(.banner-link:hover) {
-    text-decoration: underline;
   }
 
   .landing {
@@ -595,10 +414,6 @@
     max-width: 18rem;
   }
 
-  .vgc-logo {
-    max-width: 10rem;
-  }
-
   .natuurpunt-logo {
     max-width: 10rem;
   }
@@ -611,68 +426,12 @@
     background-color: var(--color-beige-light);
   }
 
-  .slow-travel-mini-festival {
-    background-color: var(--color-beige-light);
-    text-align: center;
-    padding: 8rem;
-    align-items: center;
-  }
-
-  .stmf-intro-title {
-    margin-bottom: 0;
-  }
-
-  .stmf-intro-byline {
-    text-transform: uppercase;
-    font-weight: bold;
-    margin-bottom: 3.5rem;
-    font-size: 1.8rem;
-    line-height: 1.6;
-  }
-
-  .stmf-columns {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    justify-content: space-around;
-  }
-
-  .stmf-column {
-    flex: 1 1 10%;
-    margin-left: 1rem;
-    margin-right: 1rem;
-    margin-bottom: 4rem;
-    max-width: 40rem;
-  }
-
-  .stmf-column-header {
-    background-color: var(--color-green);
-    color: var(--color-white);
-    border-radius: 0.6rem;
-    padding: 0.9rem 0.6rem;
-    text-align: center;
-    font-size: 1.8rem;
-    width: 100%;
-    margin-bottom: 1.8rem;
-  }
-
   @media only screen and (max-width: 1500px) {
     .faq-intro {
       padding: 6rem 10rem;
     }
     h1 {
       font-size: 3.4rem;
-    }
-  }
-
-  @media only screen and (max-width: 1400px) {
-    .stmf-column {
-      flex: 1 1 50%;
-    }
-
-    .stmf-column:nth-child(2),
-    .stmf-column:nth-child(3) {
-      order: 1;
     }
   }
 
@@ -763,10 +522,6 @@
     .faq-intro {
       padding: 6rem 10rem 8rem;
     }
-
-    .banner-content p {
-      font-size: 1.4rem;
-    }
   }
 
   @media only screen and (max-width: 700px) {
@@ -805,10 +560,6 @@
     .cooperation-card {
       padding: 8rem 8rem 10rem;
     }
-
-    .slow-travel-mini-festival {
-      padding: 2rem;
-    }
   }
 
   @media only screen and (max-width: 600px) {
@@ -816,18 +567,10 @@
       font-size: 2.2rem;
       line-height: 6.5rem;
     }
-
-    .stmf-intro-byline {
-      font-size: 1.6rem;
-    }
   }
 
   @media only screen and (max-width: 500px) {
     p {
-      font-size: 1.4rem;
-    }
-
-    .stmf-intro-byline {
       font-size: 1.4rem;
     }
 
@@ -853,10 +596,6 @@
 
     .cooperation-card {
       padding: 8rem 4rem 10rem;
-    }
-
-    .slow-travel-mini-festival {
-      padding: 0;
     }
   }
